+++ Readme/CHANGELOG.md
@@ -31,6 +31,7 @@
 
 - Added mTail configuration and sound alert files
 - Firewall log files are not separated for each profile (public, private, domain)
+- Set network profile for each connected hardware interface
 
 # Changelog v0.4.1